# openapi.yaml
openapi: 3.1.0
info:
  title: Traccar
  description: Traccar GPS tracking server API documentation. To use the API you need to have a server instance. For testing purposes you can use one of free [demo servers](https://www.traccar.org/demo-server/). For production use you can install your own server or get a [subscription service](https://www.traccar.org/product/tracking-server/).
  contact:
    name: Traccar Support
    url: https://www.traccar.org/
    email: support@traccar.org
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: '6.6'
servers:
  - url: https://demo.traccar.org/api
    description: Demo Server 1
  - url: https://demo2.traccar.org/api
    description: Demo Server 2
  - url: https://demo3.traccar.org/api
    description: Demo Server 3
  - url: https://demo4.traccar.org/api
    description: Demo Server 4
  - url: https://server.traccar.org/api
    description: Subscription Server
  - url: http://{host}:{port}/api
    description: Other Server
    variables:
      host:
        default: localhost
      port:
        enum:
          - '8379'
          # - '8082'
          - '80'
        default: '8379'
        # default: '8082'
security:
  - BasicAuth: []
  - ApiKey: []
...
paths:
...
  /session:
    get:
      summary: Fetch Session information
      tags:
        - Session
      parameters:
        - name: token
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Not Found
          content: {}
    post:
      summary: Create a new Session
      tags:
        - Session
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
                  format: password
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content: {}
    delete:
      summary: Close the Session
      tags:
        - Session
      responses:
        '204':
          description: No Content
          content: {}
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        readonly:
          type: boolean
        administrator:
          type: boolean
        map:
          type: string
          nullable: true
        latitude:
          type: number
        longitude:
          type: number
        zoom:
          type: integer
        password:
          type: string
        coordinateFormat:
          type: string
          nullable: true
        disabled:
          type: boolean
        expirationTime:
          type: string
          description: in IS0 8601 format. eg. `1963-11-22T18:30:00Z`
          format: date-time
          nullable: true
        deviceLimit:
          type: integer
        userLimit:
          type: integer
        deviceReadonly:
          type: boolean
        limitCommands:
          type: boolean
        fixedEmail:
          type: boolean
        poiLayer:
          type: string
          nullable: true
        attributes:
          type: object
          properties: {}
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    ApiKey:
      type: http
      scheme: bearer